Remove pytests layer of folder and shorten docstrings of each test
