+++ test/unit/test_homogenized_tangent_spd.py
@@ -3,6 +3,7 @@ import numpy as np
 import json
 import pytest
 from fans_dashboard.core.utils import identify_hierarchy, extract_and_organize_data
+from scipy.linalg import eigvalsh
 
 
 @pytest.fixture(
@@ -33,10 +34,10 @@ def test_files(request):
     pytest.skip(f"Required test files not found: {json_path} or {h5_path}")
 
 
-def test_displacement_averaging(test_files):
+def test_homogenized_tangent_spd(test_files):
     """
-    This test verifies that the average of displacement fluctuations is zero for all
-    microstructures and load cases.
+    This test verifies that the homogenized tangent is strictly Symmetric Positive Definite (SPD)
+    for all microstructures and load cases.
 
     Parameters
     ----------
@@ -44,11 +45,6 @@ def test_displacement_averaging(test_files):
         A tuple containing (input_json_file, results_h5_file) paths.
         - input_json_file: Path to the JSON file containing configuration data
         - results_h5_file: Path to the HDF5 file containing simulation results
-
-    Returns
-    -------
-    None
-        The test passes if the average of displacement fluctuations is zero.
     """
     input_json_file, results_h5_file = test_files
 
@@ -59,9 +55,11 @@ def test_displacement_averaging(test_files):
     # Check which fields are available in the results
     results = input_data.get("results", [])
 
-    # Check if displacement field is available
-    if "displacement" not in results:
-        pytest.skip(f"Skipping test: No displacement field found in {input_json_file}")
+    # Check if homogenized_tangent field is available
+    if "homogenized_tangent" not in results:
+        pytest.skip(
+            f"Skipping test: No homogenized_tangent field found in {input_json_file}"
+        )
         return
 
     # Extract hierarchy information from the h5 file
@@ -70,7 +68,7 @@ def test_displacement_averaging(test_files):
     # Load the data from the HDF5 file
     microstructures_to_load = list(hierarchy.keys())
 
-    quantities_to_load = ["displacement"]
+    quantities_to_load = ["homogenized_tangent"]
 
     time_steps_to_load = []
     load_cases_to_load = []
@@ -91,36 +89,43 @@ def test_displacement_averaging(test_files):
         time_steps_to_load,
     )
 
-    print(f"\nVerifying displacement fluctuation averages are zero...")
+    print(f"\nVerifying homogenized tangent is strictly SPD...")
 
     for microstructure in microstructures_to_load:
         for load_case in load_cases_to_load:
             if load_case in hierarchy[microstructure]:
-                if "displacement" not in data[microstructure][load_case]:
+                if "homogenized_tangent" not in data[microstructure][load_case]:
                     print(
-                        f"Skipping {microstructure}/{load_case}: Missing displacement field"
+                        f"Skipping {microstructure}/{load_case}: Missing homogenized_tangent field"
                     )
                     continue
 
-                displacement_data = data[microstructure][load_case]["displacement"]
+                tangent_data = data[microstructure][load_case]["homogenized_tangent"]
 
-                # Compute average manually by averaging over spatial dimensions (1, 2, 3)
-                # displacement_data shape: time_steps x Nx x Ny x Nz x components
-                computed_average = np.mean(displacement_data, axis=(1, 2, 3))
+                # Check each time step
+                for time_idx, tangent in enumerate(tangent_data):
+                    # Check symmetry
+                    is_symmetric = np.allclose(tangent, tangent.T, rtol=1e-5, atol=1e-8)
 
-                # Check if the computed average is approximately zero
-                zero_array = np.zeros_like(computed_average)
+                    # Check positive definiteness by computing eigenvalues
+                    eigenvalues = eigvalsh(tangent)
+                    is_positive_definite = np.all(eigenvalues > 0)
 
-                assert np.allclose(
-                    computed_average, zero_array, rtol=1e-5, atol=1e-8
-                ), (
-                    f"For microstructure {microstructure}, load case {load_case}: "
-                    f"Average of displacement fluctuations is not zero."
-                    f"\nComputed average shape: {computed_average.shape}, Value: {computed_average}"
-                )
+                    assert is_symmetric, (
+                        f"For microstructure {microstructure}, load case {load_case}, time step {time_idx}: "
+                        f"Homogenized tangent is not symmetric."
+                        f"\nTangent shape: {tangent.shape}, Max asymmetry: {np.max(np.abs(tangent - tangent.T))}"
+                    )
+
+                    assert is_positive_definite, (
+                        f"For microstructure {microstructure}, load case {load_case}, time step {time_idx}: "
+                        f"Homogenized tangent is not positive definite."
+                        f"\nEigenvalues: {eigenvalues}, Min eigenvalue: {np.min(eigenvalues)}"
+                    )
 
                 print(
-                    f"Verified: {microstructure}, load case {load_case} - displacement fluctuation average is zero"
+                    f"Verified: {microstructure}, load case {load_case} - homogenized tangent is strictly SPD "
+                    f"({tangent_data.shape[0]} time steps)"
                 )
